Escape HTML text entities in tweets
