+++ app/src/main/java/net/squanchy/tweets/service/TweetModelConverter.java
@@ -1,10 +1,5 @@
 package net.squanchy.tweets.service;
 
-import android.text.Html;
-import android.text.SpannableStringBuilder;
-import android.text.Spanned;
-import android.text.style.URLSpan;
-
 import com.google.auto.value.AutoValue;
 import com.twitter.sdk.android.core.models.HashtagEntity;
 import com.twitter.sdk.android.core.models.MediaEntity;
@@ -13,9 +8,6 @@ import com.twitter.sdk.android.core.models.Tweet;
 import com.twitter.sdk.android.core.models.UrlEntity;
 
 import java.util.List;
-import java.util.regex.MatchResult;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import net.squanchy.support.lang.Lists;
 import net.squanchy.support.lang.Optional;
@@ -25,9 +17,8 @@ import net.squanchy.tweets.domain.view.User;
 public class TweetModelConverter {
 
     private static final String MEDIA_TYPE_PHOTO = "photo";
-    private static final String BASE_TWITTER_URL = "https://twitter.com/";
-    private static final String MENTION_URL_TEMPLATE = BASE_TWITTER_URL + "%s";
-    private static final String QUERY_URL_TEMPLATE = BASE_TWITTER_URL + "search?q=%s";
+
+    private final TweetSpannedTextBuilder tweetSpannedTextBuilder = new TweetSpannedTextBuilder();
 
     TweetViewModel toViewModel(Tweet tweet) {
         User user = User.create(tweet.user.name, tweet.user.screenName, tweet.user.profileImageUrlHttps);
@@ -42,7 +33,7 @@ public class TweetModelConverter {
         return TweetViewModel.builder()
                 .id(tweet.id)
                 .text(displayableText)
-                .spannedText(applySpans(displayableText, displayTextRange.start(), hashtags, mentions, urls))
+                .spannedText(tweetSpannedTextBuilder.applySpans(displayableText, displayTextRange.start(), hashtags, mentions, urls))
                 .createdAt(tweet.createdAt)
                 .user(user)
                 .photoUrl(photoUrlMaybeFrom(photoUrls))
@@ -70,78 +61,6 @@ public class TweetModelConverter {
     private List<String> onlyPhotoUrls(List<MediaEntity> media) {
         List<MediaEntity> photos = Lists.filter(media, mediaEntity -> MEDIA_TYPE_PHOTO.equals(mediaEntity.type));
         return Lists.map(photos, mediaEntity -> mediaEntity.mediaUrlHttps);
-    }
-
-    private Spanned applySpans(String text, int startIndex, List<HashtagEntity> hashtags, List<MentionEntity> mentions, List<UrlEntity> urls) {
-        SpannableStringBuilder builder = new SpannableStringBuilder(text);
-        for (HashtagEntity hashtag : hashtags) {
-            hashtag = offsetStart(hashtag, startIndex);
-            builder.setSpan(createUrlSpanFor(hashtag), hashtag.getStart(), hashtag.getEnd(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-        }
-        for (MentionEntity mention : mentions) {
-            mention = offsetStart(mention, startIndex);
-            builder.setSpan(createUrlSpanFor(mention), mention.getStart(), mention.getEnd(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-        }
-        for (UrlEntity url : urls) {
-            url = offsetStart(url, startIndex);
-            builder.setSpan(createUrlSpanFor(url), url.getStart(), url.getEnd(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-        }
-
-        unescapeEntities(builder);
-        return builder;
-    }
-
-    private HashtagEntity offsetStart(HashtagEntity hashtag, int startIndex) {
-        return new HashtagEntity(
-                hashtag.text,
-                hashtag.getStart() - startIndex,
-                hashtag.getEnd() - startIndex
-        );
-    }
-
-    private URLSpan createUrlSpanFor(HashtagEntity hashtag) {
-        return new URLSpan(String.format(QUERY_URL_TEMPLATE, hashtag.text));
-    }
-
-    private MentionEntity offsetStart(MentionEntity mention, int startIndex) {
-        return new MentionEntity(
-                mention.id,
-                mention.idStr,
-                mention.name,
-                mention.screenName,
-                mention.getStart() - startIndex,
-                mention.getEnd() - startIndex
-        );
-    }
-
-    private URLSpan createUrlSpanFor(MentionEntity mention) {
-        return new URLSpan(String.format(MENTION_URL_TEMPLATE, mention.screenName));
-    }
-
-    private UrlEntity offsetStart(UrlEntity url, int startIndex) {
-        return new UrlEntity(
-                url.url,
-                url.expandedUrl,
-                url.displayUrl,
-                url.getStart() - startIndex,
-                url.getEnd() - startIndex
-        );
-    }
-
-    private URLSpan createUrlSpanFor(UrlEntity url) {
-        return new URLSpan(url.url);
-    }
-
-    private void unescapeEntities(SpannableStringBuilder builder) {
-        String string = builder.toString();
-        Pattern pattern = Pattern.compile("&\\w+;");
-        Matcher matcher = pattern.matcher(string);
-
-        while (matcher.find()) {
-            MatchResult matchResult = matcher.toMatchResult();
-            Spanned unescapedEntity = Html.fromHtml(matchResult.group());
-            builder.replace(matchResult.start(), matchResult.end(), unescapedEntity);
-        }
     }
 
     private Optional<String> photoUrlMaybeFrom(List<String> urls) {
